fixed withPWA error warning
